# [SPARK-40052][SQL] Handle direct byte buffers in VectorizedDeltaBinaryPackedReader