--- a/sql/core/src/test/scala/org/apache/spark/sql/execution/datasources/parquet/ParquetDeltaEncodingSuite.scala
+++ b/sql/core/src/test/scala/org/apache/spark/sql/execution/datasources/parquet/ParquetDeltaEncodingSuite.scala
@@ -246,7 +246,7 @@ abstract class ParquetDeltaEncodingSuite[T] extends ParquetCompatibilityTest
         ByteBuffer.allocate(page.length)
       }
       buf.put(page)
-      buf.rewind()
+      buf.flip()
 
       reader.initFromPage(100, ByteBufferInputStream.wrap(buf))
       readData(length, writableColumnVector, 0)