@sadikovi 2022-08-12
These are all good points. I am running the benchmarks and have updated the PR description with preliminary results. Interestingly, the results with `if` statement are slightly better which means there is a variability with the benchmark. 

@LuciferYang Do you have a specific test case that I can run to compare performance?