@LuciferYang 2022-08-12
I believe that the `unpack8Values` method needs to be executed for the Int and Long type data of Parquet v2. For the two cases we are concerned(15m data), it will be executed more than 50000 times :)