@sadikovi 2022-08-12
Thanks. @LuciferYang Do you have internal workloads that could be run with this patch to check if the performance is still on-par? I am a bit skeptical about the benchmarking results 🙂. Thanks.